requirements.txt
Click>=7.0
attrs>=19.1.0
requests>=1.25.3
tabulate>=0.8.3
